Anjay 1.10.0
Features:
- Updated avs_commons to 3.6.0, which includes:
	* an abstraction layer over allocator routines, making it possible for the
	  user to provide custom allocation/deallocation functions to be used by
	  AvsCommons,

	* removal of ``AVS_LIST_CONFIG_ALLOC/FREE`` (they are now replaced with
	  calls to ``avs_calloc()`` and ``avs_free()`` respectively),

	* removal of use of all ``time()`` calls,

	* removal of use of variable length array language feature,

	* default socket implementation refactor to use a nonblocking socket API,

	* ``avs_compat_threading`` module, implementing necessary synchronization
	  primitives used across AvsCommons such as mutexes,

	* ``avs_cleanup_global_state()`` method, allowing to (optionally) free any
	  global state implicitly instantiated in AvsCommons,

	* various compatibility fixes for FreeBSD.

- Prevented ``anjay_schedule_reconnect()`` from sending Updates when they are
  not necessary,

- Introduced an API (``anjay_get_socket_entries()``) allowing to obtain different
  kinds of sockets used by Anjay.

Improvements:
- Removed all uses of ``malloc()/calloc()/realloc()/free()`` in favor of
  AvsCommons' memory layer abstraction,
- Removed all occurrences of ``time()``,
- Removed all uses of variable length arrays,
- Improved non-GNU compilers compatibility,
- Fixed multiple typos in documentation,
- Fixed LwM2M-level error reporting for some LwM2M requests with unexpected
  payloads/content-formats (sane error codes are now returned instead of Internal
  Server Error),
- Various compatibility fixes for FreeBSD.

Fixes:
- Made bind address family depend on resolved numeric address rather than
  domain name,
- Fixed project compilation when ``WITH_AVS_PERSISTENCE`` is disabled.

+++ modules/security/src/security_transaction.c
@@ -72,7 +72,8 @@ static int validate_instance(sec_instance_t *it) {
                 it, "missing mandatory 'Short Server ID' resource value");
         return -1;
     }
-    if (_anjay_sec_validate_udp_security_mode(it->udp_security_mode)) {
+    if (_anjay_sec_validate_udp_security_mode(
+            (int32_t) it->udp_security_mode)) {
         LOG_VALIDATION_FAILED(it, "UDP Security mode %d not supported",
                               (int) it->udp_security_mode);
         return -1;
@@ -96,7 +97,8 @@ static int validate_instance(sec_instance_t *it) {
         }
     }
     if (it->has_sms_security_mode) {
-        if (_anjay_sec_validate_sms_security_mode(it->sms_security_mode)) {
+        if (_anjay_sec_validate_sms_security_mode(
+                (int32_t) it->sms_security_mode)) {
             LOG_VALIDATION_FAILED(it, "SMS Security mode %d not supported",
                                   (int) it->sms_security_mode);
             return -1;
